pyproject.toml
[tool]
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.10"
pulumi = ">=3.0"
pulumi_docker = ">=4.4.3"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
